Can now create entries from scratch.
Created a starter file from which to do from scratch entries.
Fixed bug on backend which caused the backend to crash when
similar parts could not be fetched. URIs in the sbolgraph
depend on the displayID. SeqImprove now allows you to edit
displayIDs. sbolgraph library provided no method for updating,
and protected the uris as private data. Therefore, I replace the
old display ids in the xml document with the new ones and reload
the sbol from an xml string.