Fixed 2 bugs: Invalid URI generation and annotations persisting after sequence editing.
The sbolgraph library generates invalid uris when appending sequence annotations
to the graph. My ideal solution would have been to fix the problem within the
library, but I don't have the ability to edit the library. Also, the sbolgraph
library encapsulates data in such a way that I can't directly modify the graph, I
have to use accessor methods which are riddled with bugs. So, my inelegant solution
is to serialize the sbol document to xml and fix the invalid uris in the xml. Then,
I reload the sbol document from the fixed xml.

For the second bug, I just had to call setActive on all active annotations, and
fix a react bug with duplicate keys on elements so that the annotations section
would rerender.