Text description edit bug fixes
Previously, editing a text description would screw up the
placement of the annotations. That is fixed.

Previusly, removing the entire text description would
cause all changes to be discarded and take the user
out of edit mode. That is fixed.

Also, now, when all mentions of a text annotatino are
removed from the text description, the annotation is
deleted.