Booting API server in advance & UI bug fix
In the previous commit, I wrote some code
that would tell the server to basically load
in some data, the feature library data, into
computer memory after the first request to
the server, so it can just use it for every
request afterward. This is good, because
loading that data takes a lot of time. But,
what I realized when deploying the code is
that Azure doesn't just keep our server
running all the time. Without any requests
being made to the server, it gets shut off.
This means that it has to be rebooted, and
that initialization process has to start
all over again. This nullifies the time
savings of the previous update because
there aren't enough people using SeqImprove
to keep the server running. So, my solution
is to create an API endpoint whose only purpose
is to boot up the server and trigger the setup
function to load our feature library data,
if and only if the server is just starting up.
Then I make the front end request this API
endpoint any time SeqImrpove is visited.
This gets the process of loading in the part
libraries started before the user even begins
to upload an sbol file, let alone try to
analyze the DNA.

I also fixed a little UI bug that my last
commit unfortunately created. The tooltip
for displaying the part library somehow
prevented the checkbox for sequence annotations
from being unclicked.